backend/get_recipes/requirements.txt
boto3==1.34.146
botocore==1.34.146
jmespath==1.0.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.2
